+++ CHANGELOG.md
@@ -5,6 +5,30 @@ Toutes les modifications notables de ce projet seront documentées dans ce fichi
 Le format est basé sur [Keep a Changelog](https://keepachangelog.com/fr/1.0.0/),
 et ce projet adhère au [Semantic Versioning](https://semver.org/lang/fr/).
 
+<<<<<<< HEAD
+=======
+## [5.1.1] - 2026-06-24
+
+### Ajouté
+
+### Modifié
+- Modification email metadata
+
+## [5.1.0] - 2026-06-24
+
+### Ajouté
+- Documentation développeur (#178)
+
+### Modifié
+- Optimisation de la synchronisation SQLite (vacuum, numéro de réconciliation)
+- Amélioration de la gestion des contraintes bloquantes : validation REGEX avant commit pour éviter l'affichage inutile du bandeau d'erreur (#187)
+- Amélioration de la gestion des transactions et des erreurs proxy
+- Réduction du délai d'attente lors de la validation serveur
+- gestion des erreurs de connexion améliorée
+- Mise à jour de la documentation
+
+
+>>>>>>> community
 ## [5.0.9] - 2026-05-28
 
 ### Modifié